Split up into indexer and kb_builder middleware
